Save changes

+++ backend/controllers/jobController.js
@@ -4,54 +4,54 @@ const Job = require('../models/jobModel')
 
 const getJobs = asyncHandler(async (req, res) => {
   try {
-    const { page = 1, limit = 10, title, company, location, sort } = req.query;
+    const { page = 1, limit = 10, position, company, location, sort } = req.query;
     const skip = (page - 1) * limit;
 
     let query = {};
 
-    if (title) {
-      query.position = { $regex: new RegExp(title, 'i') };
+    if (position) {
+      query.position = { $regex: new RegExp(position, 'i') };
     }
 
-  if (company) {
+    if (company) {
     query.company = { $regex: new RegExp(company, 'i') };
-  }
-
-  if (location) {
-    query.location = { $regex: new RegExp(location, 'i') };
-  }
-
-  const sortOptions = {};
-
-  if (sort) {
-    switch (sort) {
-      case 'title':
-        sortOptions.position = 1;
-        break;
-      case 'company':
-        sortOptions.company = 1;
-        break;
-      case 'location':
-        sortOptions.location = 1;
-        break;
-      case 'postedAt':
-        sortOptions.postedAt = -1;
-        break;
-      default:
-        sortOptions.postedAt = -1;
-        break;
     }
-  }
 
-  const jobs = await Job.find(query)
-    .skip(skip)
-    .limit(Number(limit))
-    .sort(sortOptions);
+    if (location) {
+    query.location = { $regex: new RegExp(location, 'i') };
+    }
 
-  res.json(jobs);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
+    const sortOptions = {};
+
+    if (sort) {
+      switch (sort) {
+        case 'position':
+          sortOptions.position = 1;
+          break;
+        case 'company':
+          sortOptions.company = 1;
+          break;
+        case 'location':
+          sortOptions.location = 1;
+          break;
+        case 'postedAt':
+          sortOptions.postedAt = -1;
+          break;
+        default:
+          sortOptions.postedAt = -1;
+          break;
+      }
+    }
+
+    const jobs = await Job.find(query)
+      .skip(skip)
+      .limit(Number(limit))
+      .sort(sortOptions);
+
+    res.json(jobs);
+    } catch (err) {
+      res.status(500).json({ error: err.message });
+    }
 })
 
 const setJob = asyncHandler(async (req, res) => {
@@ -75,8 +75,14 @@ const updateJob = asyncHandler(async (req, res) => {
 
 const deleteJob = asyncHandler(async (req, res) => {
   try {
-    const job = await Job.findByIdAndRemove(req.params.id);
-    res.status(200).json({id: req.params.id, jobdata: job});
+    const job = await Job.findById(req.params.id)
+
+    if(!job) {
+      res.status(400)
+      throw new Error('Job not Found')
+    }
+    await Job.findByIdAndRemove(req.params.id);
+    res.status(200).json({id: req.params.id});
   } catch (err) {
     res.status(400).json({ error: err.message });
   }
